Update it.json
I was working on a timeline and got aware that Italian translation was quite inaccurate. So I did my best :)
I'm still in doubt about "Eras", now "Epoche", that in my job (teaching Italian SL) is too big and/or vague. Maybe, better using "period of time" or something like that. I'm gonna think over and share with my colleagues.
Anyway, it's much better, I believe, even if's not perfect. We can progress...